refactor: DefaultEvidenceUpdater to new file for easier code review
